Parse nested expressions, instead of only ArithExpressions
Heavy updates to the parser.
Now all expressions can be chained with + and minus (as they currently all evaluate to ints).
let a = 2 in print a is now possible.
So is `let a = 2 in let b = 3 in let c = a + b + 3 - 4 in let res = a + b + c in print res`